src/requirements.txt
aiohttp==3.9.5 ; python_version >= "3.8" and python_version < "4.0"
aiosignal==1.3.1 ; python_version >= "3.8" and python_version < "4.0"
aiounittest==1.3.0 ; python_version >= "3.8" and python_version < "4.0"
annotated-types==0.7.0 ; python_version >= "3.8" and python_version < "4.0"
anyio==4.4.0 ; python_version >= "3.8" and python_version < "4.0"
async-timeout==4.0.3 ; python_version >= "3.8" and python_version < "3.11"
attrs==23.2.0 ; python_version >= "3.8" and python_version < "4.0"
azure-ai-contentsafety==1.0.0 ; python_version >= "3.8" and python_version < "4.0"
azure-core==1.30.2 ; python_version >= "3.8" and python_version < "4.0"
azure-identity==1.17.1 ; python_version >= "3.8" and python_version < "4.0"
babel==2.9.1 ; python_version >= "3.8" and python_version < "4.0"
botbuilder-core==4.16.0 ; python_version >= "3.8" and python_version < "4.0"
botbuilder-dialogs==4.16.0 ; python_version >= "3.8" and python_version < "4.0"
botbuilder-integration-aiohttp==4.16.0 ; python_version >= "3.8" and python_version < "4.0"
botbuilder-schema==4.16.0 ; python_version >= "3.8" and python_version < "4.0"
botframework-connector==4.16.0 ; python_version >= "3.8" and python_version < "4.0"
botframework-streaming==4.16.0 ; python_version >= "3.8" and python_version < "4.0"
certifi==2024.7.4 ; python_version >= "3.8" and python_version < "4.0"
cffi==1.16.0 ; python_version >= "3.8" and python_version < "4.0" and platform_python_implementation != "PyPy"
charset-normalizer==3.3.2 ; python_version >= "3.8" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.8" and python_version < "4.0" and platform_system == "Windows"
cryptography==42.0.8 ; python_version >= "3.8" and python_version < "4.0"
dataclasses-json==0.6.7 ; python_version >= "3.8" and python_version < "4.0"
datedelta==1.4 ; python_version >= "3.8" and python_version < "4.0"
distro==1.9.0 ; python_version >= "3.8" and python_version < "4.0"
emoji==1.7.0 ; python_version >= "3.8" and python_version < "4.0"
exceptiongroup==1.2.1 ; python_version >= "3.8" and python_version < "3.11"
frozenlist==1.4.1 ; python_version >= "3.8" and python_version < "4.0"
grapheme==0.6.0 ; python_version >= "3.8" and python_version < "4.0"
h11==0.14.0 ; python_version >= "3.8" and python_version < "4.0"
httpcore==1.0.5 ; python_version >= "3.8" and python_version < "4.0"
httpx==0.27.0 ; python_version >= "3.8" and python_version < "4.0"
idna==3.7 ; python_version >= "3.8" and python_version < "4.0"
importlib-resources==6.4.0 ; python_version >= "3.8" and python_version < "3.9"
isodate==0.6.1 ; python_version >= "3.8" and python_version < "4.0"
jsonpickle==1.4.2 ; python_version >= "3.8" and python_version < "4.0"
jsonschema-specifications==2023.12.1 ; python_version >= "3.8" and python_version < "4.0"
jsonschema==4.22.0 ; python_version >= "3.8" and python_version < "4.0"
marshmallow==3.21.3 ; python_version >= "3.8" and python_version < "4.0"
msal-extensions==1.2.0 ; python_version >= "3.8" and python_version < "4.0"
msal==1.29.0 ; python_version >= "3.8" and python_version < "4.0"
msrest==0.7.1 ; python_version >= "3.8" and python_version < "4.0"
multidict==6.0.5 ; python_version >= "3.8" and python_version < "4.0"
multipledispatch==1.0.0 ; python_version >= "3.8" and python_version < "4.0"
mypy-extensions==1.0.0 ; python_version >= "3.8" and python_version < "4.0"
oauthlib==3.2.2 ; python_version >= "3.8" and python_version < "4.0"
openai==1.35.10 ; python_version >= "3.8" and python_version < "4.0"
packaging==24.1 ; python_version >= "3.8" and python_version < "4.0"
pkgutil-resolve-name==1.3.10 ; python_version >= "3.8" and python_version < "3.9"
portalocker==2.10.0 ; python_version >= "3.8" and python_version < "4.0"
pycparser==2.22 ; python_version >= "3.8" and python_version < "4.0" and platform_python_implementation != "PyPy"
pydantic-core==2.20.1 ; python_version >= "3.8" and python_version < "4.0"
pydantic==2.8.2 ; python_version >= "3.8" and python_version < "4.0"
pyjwt==2.8.0 ; python_version >= "3.8" and python_version < "4.0"
pyjwt[crypto]==2.8.0 ; python_version >= "3.8" and python_version < "4.0"
python-dotenv==1.0.1 ; python_version >= "3.8" and python_version < "4.0"
pytz==2024.1 ; python_version >= "3.8" and python_version < "4.0"
pywin32==306 ; python_version >= "3.8" and python_version < "4.0" and platform_system == "Windows"
pyyaml==6.0.1 ; python_version >= "3.8" and python_version < "4.0"
recognizers-text-choice==1.0.2a2 ; python_version >= "3.8" and python_version < "4.0"
recognizers-text-date-time==1.0.2a2 ; python_version >= "3.8" and python_version < "4.0"
recognizers-text-number-with-unit==1.0.2a2 ; python_version >= "3.8" and python_version < "4.0"
recognizers-text-number==1.0.2a2 ; python_version >= "3.8" and python_version < "4.0"
recognizers-text==1.0.2a2 ; python_version >= "3.8" and python_version < "4.0"
referencing==0.35.1 ; python_version >= "3.8" and python_version < "4.0"
regex==2024.5.15 ; python_version >= "3.8" and python_version < "4.0"
requests-oauthlib==2.0.0 ; python_version >= "3.8" and python_version < "4.0"
requests==2.32.3 ; python_version >= "3.8" and python_version < "4.0"
rpds-py==0.18.1 ; python_version >= "3.8" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.8" and python_version < "4.0"
sniffio==1.3.1 ; python_version >= "3.8" and python_version < "4.0"
git+https://github.com/microsoft/teams-ai.git@main#subdirectory=python/packages/ai ; python_version >= "3.8" and python_version < "4.0"
tiktoken==0.7.0 ; python_version >= "3.8" and python_version < "4.0"
tqdm==4.66.4 ; python_version >= "3.8" and python_version < "4.0"
types-pyyaml==6.0.12.20240311 ; python_version >= "3.8" and python_version < "4.0"
typing-extensions==4.12.2 ; python_version >= "3.8" and python_version < "4.0"
typing-inspect==0.9.0 ; python_version >= "3.8" and python_version < "4.0"
urllib3==1.26.19 ; python_version >= "3.8" and python_version < "4.0"
yarl==1.9.4 ; python_version >= "3.8" and python_version < "4.0"
zipp==3.19.2 ; python_version >= "3.8" and python_version < "3.9"
